# api_doc.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Home","operationId":"home__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/user/create/":{"post":{"summary":"Create User","operationId":"create_user_user_create__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBase"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/all/":{"get":{"summary":"All Users List","operationId":"all_users_list_user_all__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/user/{user_id}/":{"get":{"summary":"Get User By Id","operationId":"get_user_by_id_user__user_id___get","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","title":"User Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/book/create/":{"post":{"summary":"Create Book","operationId":"create_book_book_create__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookBase"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/book/all/":{"get":{"summary":"All Book List","operationId":"all_book_list_book_all__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/book/{book_id}/":{"get":{"summary":"Get Book By Id","operationId":"get_book_by_id_book__book_id___get","parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/book/{book_id}/details/":{"put":{"summary":"Update Book Details","operationId":"update_book_details_book__book_id__details__put","parameters":[{"name":"book_id","in":"path","required":true,"schema":{"type":"integer","title":"Book Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookDetailsUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/borrowed-books/borrow/":{"post":{"summary":"Borrow Book","operationId":"borrow_book_borrowed_books_borrow__post","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}},{"name":"book_id","in":"query","required":true,"schema":{"type":"integer","title":"Book Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/borrowed-books/return/":{"put":{"summary":"Return Book","operationId":"return_book_borrowed_books_return__put","parameters":[{"name":"user_id","in":"query","required":true,"schema":{"type":"integer","title":"User Id"}},{"name":"book_id","in":"query","required":true,"schema":{"type":"integer","title":"Book Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/borrowed-books/all/":{"get":{"summary":"List All Borrowed Books","operationId":"list_all_borrowed_books_borrowed_books_all__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"BookBase":{"properties":{"title":{"type":"string","title":"Title"},"isbn":{"type":"string","title":"Isbn"},"published_date":{"type":"string","title":"Published Date"},"genre":{"type":"string","title":"Genre"}},"type":"object","required":["title","isbn","published_date","genre"],"title":"BookBase"},"BookDetailsUpdate":{"properties":{"number_of_pages":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Number Of Pages"},"publisher":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Publisher"},"language":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Language"}},"type":"object","required":["number_of_pages","publisher","language"],"title":"BookDetailsUpdate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"UserBase":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","title":"Email"},"membership_date":{"type":"string","title":"Membership Date"}},"type":"object","required":["name","email","membership_date"],"title":"UserBase"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}